DeepEval/requirements.txt
deepeval
python-dotenv
llama-index